+++ app/rdev/framework/http/middleware/Session.php
@@ -6,27 +6,30 @@
  */
 namespace RDev\Framework\HTTP\Middleware;
 use Closure;
-use DateTime;
+use RDev\Applications\Paths;
 use RDev\HTTP\Middleware\IMiddleware;
 use RDev\HTTP\Requests\Request;
-use RDev\HTTP\Responses\Cookie;
 use RDev\HTTP\Responses\Response;
 use RDev\Sessions\ISession;
 use SessionHandlerInterface;
 
-class Session implements IMiddleware
+abstract class Session implements IMiddleware
 {
+    /** @var Paths The application paths */
+    protected $paths = null;
     /** @var ISession The session used by the application */
     protected $session = null;
     /** @var SessionHandlerInterface The session handler used by the application */
     protected $sessionHandler = null;
 
     /**
+     * @param Paths $paths The application paths
      * @param ISession $session The session used by the application
      * @param SessionHandlerInterface $sessionHandler The session handler used by the application
      */
-    public function __construct(ISession $session, SessionHandlerInterface $sessionHandler)
+    public function __construct(Paths $paths, ISession $session, SessionHandlerInterface $sessionHandler)
     {
+        $this->paths = $paths;
         $this->session = $session;
         $this->sessionHandler = $sessionHandler;
     }
@@ -39,7 +42,21 @@ class Session implements IMiddleware
         $this->startSession($request);
         $response = $next($request);
         $this->writeSession($response);
+
+        return $response;
     }
+
+    /**
+     * Runs garbage collection, if necessary
+     */
+    abstract protected function gc();
+
+    /**
+     * Writes any session data needed in the response
+     *
+     * @param Response $response The response to write to
+     */
+    abstract protected function writeToResponse(Response $response);
 
     /**
      * Starts the session
@@ -48,10 +65,17 @@ class Session implements IMiddleware
      */
     protected function startSession(Request $request)
     {
+        $this->gc();
         $this->session->setId($request->getCookies()->get($this->session->getName()));
-        $this->sessionHandler->open(/* TODO: IMPLEMENT */
-            null, $this->session->getId());
-        $this->session->start($this->sessionHandler->read($this->session->getId()));
+        $this->sessionHandler->open(null, $this->session->getName());
+        $sessionVariables = @unserialize($this->sessionHandler->read($this->session->getId()));
+
+        if($sessionVariables === false)
+        {
+            $sessionVariables = [];
+        }
+
+        $this->session->start($sessionVariables);
     }
 
     /**
@@ -63,18 +87,6 @@ class Session implements IMiddleware
     {
         $this->session->ageFlashData();
         $this->sessionHandler->write($this->session->getId(), serialize($this->session->getAll()));
-        $sessionCookie = new Cookie(
-            $this->session->getName(),
-            $this->session->getId(),
-            /* TODO: IMPLEMENT */
-            new DateTime(),
-            /* TODO: IMPLEMENT */
-            "",
-            /* TODO: IMPLEMENT */
-            "",
-            /* TODO: IMPLEMENT */
-            false
-        );
-        $response->getHeaders()->setCookie($sessionCookie);
+        $this->writeToResponse($response);
     }
 }