+++ app/rdev/sessions/ids/IdGenerator.php
@@ -11,7 +11,6 @@ class IdGenerator implements IIdGenerator
 {
     /** The default length of an Id */
     const DEFAULT_LENGTH = 40;
-
     /** @var Strings The strings utility */
     private $strings = null;
 
@@ -29,5 +28,19 @@ class IdGenerator implements IIdGenerator
     public function generate($length = self::DEFAULT_LENGTH)
     {
         return $this->strings->generateRandomString($length);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isIdValid($id)
+    {
+        $regex = sprintf(
+            "/^[a-z0-9]{%d,%d}$/i",
+            self::MIN_LENGTH,
+            self::MAX_LENGTH
+        );
+
+        return is_string($id) && preg_match($regex, $id);
     }
 }